# src/main/resources/ecommers-swagger-openapi-code-generator.yaml
openapi: 3.0.3
info:
  title: Simple E-commerce API
  description: Backend API for product listing, cart, orders, and user authentication.
  version: 1.0.0
  contact:
    email: sreenivasakurabalakota@gmail.com
    name: Kurabalakota Sreenivasa

servers:
  - url: http://localhost:7896

tags:
  - name: login
    description: Login forms to login with details
  - name: Users
    description: used to  register records
  - name: Products
    description: used to deal with products
  - name: Cart
    description: Shopping cart operations
  - name: Orders
    description: Order placement and history
  - name : Admin-User
    description: user end points accessed by admin
paths:

  /login:
    post:
      summary: Login a user
      tags: [login]
      operationId: userLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login An User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'


  /user:

    post:
      summary: Create a new user
      tags:
        - Users
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input

    get:
      summary: get all users
      tags:
        - Users
      operationId: getAllUsers
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input


  /user/{email}:
    get:
      summary: Get an user by email
      tags:
        - Users
      operationId: getUserByEmail
      parameters:
        - in: path
          name: email
          description: email of the user
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User Fetched Successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found

  /user/{id}:
    get:
      summary: Get an user by ID
      tags:
        - Users
      operationId: getUserById
      parameters:
        - in: path
          name: id
          description: Id of the user
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User Fetched Successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found

    patch:
      summary: update an user by ID
      tags:
        - Users
      operationId: updateUserById
      parameters:
        - in: path
          name: id
          description: ID of the user to update
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found

    delete:
      tags:
        - Users
      operationId: deleteUserById
      summary: Delete an user by ID
      parameters:
        - in: path
          name: id
          description: ID of user to delete
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: string
        '404':
          description: User not found

  /user/{id}/password:

    patch:
      summary: to update a password
      tags:
        - Users
      operationId: toUpdatePassword
      parameters:
        - in: path
          name: id
          description: User Id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                oldPassword:
                  type: string
                newPassword:
                  type: string
      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Incorrect old password
          content:
            application/json:
              schema:
                type: string
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: string

  /products:
    get:
      tags:
        - Products
      summary: Get all products
      operationId: getAllProducts
      responses:
        '200':
          description: Fetched all products successsfully....!
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      tags:
        - Products
      summary: create a Product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Created product successfully....!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'

  /products/{id}:
    get:
      tags:
        - Products
      summary: Get products by id
      operationId: getProductById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Fetched a product successfully...!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        default:
          description: Failed to fetch Product

    put:
      tags:
        - Products
      summary:  Update a product by Id
      operationId: updateProductById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Updated a product successfully...!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        default:
          description: Failed to update Product

    delete:
      tags:
        - Products
      summary:  Delete a product by Id
      operationId: deleteProductById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Deleted a product successfully...!
          content:
            application/json:
              schema:
                type: string
        default:
          description: Failed to delete Product

  /cart/{userId}:
    get:
      tags:
        - Cart
      summary: Get cart of User by Id
      operationId: getCartByUserId
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User's Cart with items fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '404':
          description: Cart not found

    delete:
      tags:
        - Cart
      summary: Clear all items from the user's cart
      operationId: clearCartByUserId
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Cart cleared successfully
          content:
            application/json:
              schema:
                type: string

  /cart/items:
    post:
      tags:
        - Cart
      summary: Add item to cart
      operationId: addItemToCart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItemRequest'
      responses:
        '201':
          description: Item Added to Cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartItem'

  /cart/items/{itemId}:

    put:
      tags:
        - Cart
      summary: Update quantity of an item in the cart
      operationId: updateCartItemQuantity
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
                  minimum: 1

      responses:
        '200':
          description: Cart item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartItem'

        '404':
          description: Cart item not found

    delete:
      tags:
        - Cart
      summary: Remove an item from the cart
      operationId: removeCartItem
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Cart item removed successfully
          content:
            application/json:
              schema:
                type: string
        '404':
          description:  Cart item not found

  /orders:
    post:
      tags:
        - Orders
      summary: place an order from user cart
      operationId: placeOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
      responses:
        '200':
          description: Order placed successfully
          content:
            applicaton/json:
              schema:
                $ref: '#/components/schemas/Order'
        default:
          description: faild ordering

  /orders/user/{userId}:
    get:
      tags:
        - Orders
      summary: get all orders of an User
      operationId: getAllOrdersByUser
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Orders fetched successfully...!
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'

  /orders/{orderId}:
    get:
      tags:
        - Orders
      summary: get a single order by Id
      operationId: getOrderById
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order fetched successfully..!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

components:

  schemas:

    LoginRequest:
      type: object
      required: [email,password]
      properties:
        email:
          type: string
          example: xyz@gmail.com
        password:
          type: string
          example: JohnWick@1234

    LoginResponse:
      type: object
      required: [type,token]
      properties:
        type:
          type: string
          example: Bearer
        token:
          type: string

    UserCreateRequest:
      type: object
      required: [name,password,email,phoneNumber,address]
      properties:
        id:
          type: integer
          example: 1234
          readOnly: true
        name:
          type: string
          example: john
          minLength: 3
          maxLength: 25
          x-message: "Enter a valid name"
        password:
          type: string
          example: John@123
          pattern: "^(?=.*[A-Z])(?=.*[a-z])(?=.*\\d)(?=.*[@#$%^&+=!]).*$"
          x-message:  "Password must contain Uppercase, Lowercase, Digit, Special character"
          minLength: 8
          maxLength: 14
        role:
          type: string
          enum: [CUSTOMER, ADMIN]
          readOnly: true
        email:
          type: string
          format: email
          x-message: "Enter a valid email address"
        createdDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$'
          readOnly: true
        updatedAt:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$'
          readOnly: true
        phoneNumber:
          type: string
          pattern: ^[0-9]{10}$
          x-message: "Enter a valid phone number"
        address:
          type: string
          minLength: 3
          maxLength: 120
          x-message: "Enter a valid Address"

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1234
        name:
          type: string
          example: John Wick
        password:
          type: string
          example: JohnWick@1234
        role:
          type: string
          enum: [CUSTOMER, ADMIN]
        email:
          type: string
          example: xyz@gmail.com
        createdDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$'
        updatedAt:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$'
        phoneNumber:
          type: string
        address:
          type: string

    User:
      type: object
      properties:
        id:
          type: integer
          example: 1234
        name:
          type: string
          example: John Wick
        password:
          type: string
          example: JohnWick@1234
        role:
          type: string
          enum: [CUSTOMER, ADMIN]
        email:
          type: string
          example: xyz@gmail.com
        createdDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$'
        updatedAt:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$'
        phoneNumber:
          type: string
        address:
          type: string

    UserUpdateRequest:
      type: object
      properties:
        name:
          type: string
        password:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        address:
          type: string

    Product:
      type: object
      properties:
        id:
          type: integer
          example: 4567
        name:
          type: string
          example: mobiles
        description:
          type: string
        price:
          type: number
          format: double
        createdAt:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$'
        updatedAt :
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}$'
        stockQuantity:
          type: integer

    Cart:
      type: object
      properties:
        cartItems:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        totalAmount:
          type: number
          format: double

    CartItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer
        total:
          type: number

    CartItemRequest:
      type: object
      required: [productId, quantity,userId]
      properties:
        userId:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer

    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        name:
          type: string
        quantity:
          type: integer
        price:
          type: number
          format: double
        total:
          type: number
          format: double

    Order:
      type: object
      properties:
        orderId:
          type: integer
        orderDate:
          type: string
          format: date-time
          x-java-type: java.time.LocalDateTime
        totalAmount:
          type: number
          format: double
        status:
          type: string
          example: PENDING
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'



